Promote Utils.getErrorMessage and StringBuilder.plusAssign to jvmCommonMain for future use in jvmCli
