+++ src/jvmCommonMain/kotlin/au/id/micolous/metrodroid/util/UtilsJvm.kt
@@ -1,0 +1,36 @@
+/*
+ * UtilsJvm.kt
+ *
+ * Copyright 2011 Eric Butler <eric@codebutler.com>
+ * Copyright 2015-2019 Michael Farrell <micolous+git@gmail.com>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package au.id.micolous.metrodroid.util
+
+operator fun StringBuilder.plusAssign(other: String) {
+    this.append(other)
+}
+
+/**
+ * Gets an error message for a [Throwable], preferring a localized message if available.
+ */
+fun getErrorMessage(ex: Throwable?): String {
+    if (ex == null)
+        return "unknown error"
+    val ex = ex.cause ?: ex
+    val errorMessage = ex.localizedMessage?.ifEmpty { null } ?: ex.message?.ifEmpty { null } ?: ex.toString()
+    return ex.javaClass.simpleName + ": " + errorMessage
+}
